After merge remove unnecessary casts
(uint32_t)IFINITE

+++ contribs/perlapi/common/msg.h
@@ -45,7 +45,7 @@ inline static int av_store_uint32_t(AV* av, int index, uint32_t val)
 	/* Perl has a hard time figuring out the an unsigned int is
 	   equal to INFINITE or NO_VAL since they are treated as
 	   signed ints so we will handle this here. */
-	if(val == (uint32_t)INFINITE)
+	if(val == INFINITE)
 		sv = newSViv(INFINITE);
 	else if(val == NO_VAL)
 		sv = newSViv(NO_VAL);
@@ -131,7 +131,7 @@ inline static int hv_store_uint32_t(HV* hv, const char *key, uint32_t val)
 	/* Perl has a hard time figuring out the an unsigned int is
 	   equal to INFINITE or NO_VAL since they are treated as
 	   signed ints so we will handle this here. */
-	if(val == (uint32_t)INFINITE)
+	if(val == INFINITE)
 		sv = newSViv(INFINITE);
 	else if(val == NO_VAL)
 		sv = newSViv(NO_VAL);
